+++ tests/test_rma.py
@@ -3,12 +3,15 @@ import pytest
 import ucp
 
 
+@pytest.mark.asyncio
 @pytest.mark.parametrize("blocking_progress_mode", [True, False])
-def test_fence(blocking_progress_mode):
+async def test_fence(blocking_progress_mode):
+    # Test needs to be async here to ensure progress tasks are cleared
+    # and avoid warnings.
+
     ucp.init(blocking_progress_mode=blocking_progress_mode)
     # this should always succeed
     ucp.fence()
-    ucp.reset()
 
 
 @pytest.mark.asyncio
@@ -17,4 +20,3 @@ async def test_flush(blocking_progress_mode):
     ucp.init(blocking_progress_mode=blocking_progress_mode)
 
     await ucp.flush()
-    ucp.reset()
