+++ tests/test_send_recv.py
@@ -1,4 +1,3 @@
-import asyncio
 import functools
 
 import pytest
@@ -9,24 +8,6 @@ np = pytest.importorskip("numpy")
 
 msg_sizes = [2 ** i for i in range(0, 25, 4)]
 dtypes = ["|u1", "<i8", "f8"]
-
-
-def handle_exception(loop, context):
-    msg = context.get("exception", context["message"])
-    print(msg)
-
-
-# Let's make sure that UCX gets time to cancel
-# progress tasks before closing the event loop.
-@pytest.fixture()
-def event_loop(scope="function"):
-    loop = asyncio.new_event_loop()
-    loop.set_exception_handler(handle_exception)
-    ucp.reset()
-    yield loop
-    ucp.reset()
-    loop.run_until_complete(asyncio.sleep(0))
-    loop.close()
 
 
 def make_echo_server(create_empty_data):
